initialize firebase
