+++ src/config/DataContext.tsx
@@ -1,0 +1,8 @@
+import { createContext } from "react";
+import { DataContextInterface } from "./Interface";
+
+const DefaultDataContext = {
+    
+}
+
+export const DataContext = createContext<DataContextInterface>(DefaultDataContext)
