+++ src/pages/johe/Johe.tsx
@@ -1,5 +1,6 @@
 import Navigation from "../../components/navigation/Navigation";
 import HomePageAbstract from "./HomePageAbstract/HomePageAbstract";
+import LikeButton from "./LikeButton/LikeButton";
 import ListCard from "./ListCard/ListCard";
 import ListPageAbstract from "./ListPageAbstract/ListPageAbstract";
 
@@ -10,6 +11,7 @@ export default function Johe() {
       <ListCard/>
       <HomePageAbstract/>
       <ListPageAbstract/>
+      <LikeButton/>
     </>
   )
 }
