+++ src/pages/johe/LikeButton/LikeButton.tsx
@@ -1,0 +1,7 @@
+export default function LikeButton() {
+    return (
+        <div className="fixed right-3 bottom-3 w-11 h-11 flex justify-center items-center shadow-lg p-3 bg-yellow rounded-full hover:opacity-80 cursor-pointer">
+            <i className="fa-regular fa-heart"></i>
+        </div>
+    )
+}
